requirements.txt
pillow
pandas
numpy
ipykernel
waitress
flask
load-dotenv
xlwt
python-dotenv 